+++ lib/lightning/accounts/user_notifier.ex
@@ -350,6 +350,20 @@ defmodule Lightning.Accounts.UserNotifier do
     """)
   end
 
+  def deliver_project_invitation_email(user, inviter, project, role, token) do
+    deliver(user.email, "Join #{project.name} on OpenFn as a collaborator", """
+    Hi #{user.first_name},
+
+    #{inviter.first_name} has invited you to join project "#{project.name}" and granted you "#{role}" access. Since you don't have an OpenFn account yet, we've set one up for you.
+
+    Please click the link below to complete your account setup: #{url(LightningWeb.Endpoint, ~p"/users/reset_password/#{token}")}
+
+    If you did not request to join this project, please ignore this email.
+
+    OpenFn
+    """)
+  end
+
   defp pluralize_with_s(1, string), do: string
   defp pluralize_with_s(_integer, string), do: "#{string}s"
 end