+++ lib/lightning_web/live/project_live/collaborators.ex
@@ -43,27 +43,38 @@ defmodule LightningWeb.ProjectLive.Collaborators do
 
         emails = Enum.map(collaborators, &get_field(&1, :email))
 
-        existing_users = Lightning.Accounts.list_users_by_emails(emails)
+        existing_users =
+          Lightning.Accounts.list_users_by_emails(emails)
+
+        existing_emails =
+          Enum.map(existing_users, fn user -> String.downcase(user.email) end)
+
+        {existing_collaborators, new_collaborators} =
+          Enum.split_with(collaborators, fn collaborator ->
+            Enum.member?(
+              existing_emails,
+              get_field(collaborator, :email) |> String.downcase()
+            )
+          end)
 
         updated_collaborators =
           validate_collaborators(
-            collaborators,
+            existing_collaborators,
             existing_users,
             current_project_users
           )
 
-        put_embed(changeset, :collaborators, updated_collaborators)
+        put_embed(
+          changeset,
+          :collaborators,
+          updated_collaborators ++ new_collaborators
+        )
       else
         changeset
       end
 
     with {:ok, %{collaborators: collaborators}} <-
            apply_action(changeset, :insert) do
-      collaborators =
-        Enum.map(collaborators, fn c ->
-          Map.take(c, [:user_id, :role])
-        end)
-
       {:ok, collaborators}
     end
   end
@@ -76,7 +87,8 @@ defmodule LightningWeb.ProjectLive.Collaborators do
     Enum.map(collaborators, fn collaborator ->
       existing_user =
         Enum.find(existing_users, fn u ->
-          u.email == get_field(collaborator, :email)
+          String.downcase(u.email) ==
+            get_field(collaborator, :email) |> String.downcase()
         end)
 
       collaborator
@@ -84,7 +96,7 @@ defmodule LightningWeb.ProjectLive.Collaborators do
       |> validate_change(:email, fn :email, _email ->
         cond do
           is_nil(existing_user) ->
-            [email: "There is no account connected this email"]
+            []
 
           Enum.find(current_project_users, &(&1.user_id == existing_user.id)) ->
             [email: "This account is already part of this project"]